requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.133.0
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.2.10
langchain-core==1.2.15
langchain-openai==1.1.10
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.17.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
openai==2.23.0
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
plotly==6.5.2
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
pypdfium2==5.5.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tifffile==2026.2.20
tiktoken==0.12.0
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
xxhash==3.6.0
zstandard==0.25.0
